LabBreakers/sumalabs/src/requirements.txt
about-time==3.1.1
alive-progress==2.1.0
argcomplete==3.0.3
grapheme==0.6.0
importlib-metadata==4.8.3
Nuitka==1.5.3
orderedset==2.0.3
PyYAML==6.0
typing_extensions==4.1.1
zipp==3.6.0
zstandard==0.20.0
